All the unit tests still pass, though there may be some hidden issues from the conversion
* Converted many things from u32 to s32
* Converted a few things from u32 to size_t (for non-32-bit portability)

+++ src/utilities.cpp
@@ -50,7 +50,7 @@ namespace Anki
 
   u64 Log2(u64 x)
   {
-    u32 powerCount = 0;
+    u64 powerCount = 0;
     // While x is even and greater than 1, keep dividing by two
     while (x >>= 1) {
       powerCount++;
